switch to username@timestamp.pub key names
this allows to exploit a feature of gitolite, in which gitolite
will construct the username of the gitolite user only of the part
leading up to the @. this makes the gitolite usernames much easier
to predict.

also removes the "redmine_" prefix.

this change might with existing databases, as the format of
the gitolite_public_key identifier changes. I had to remove all keys
from the database and re-add them once until gitolite worked again
correctly.